🔧 [CMN] BFF 미들웨어 서버 구성
